requirements.txt
## WARNING: Replaced by Pipfile

# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
